requirements.txt
pyrogram==2.0.106
tgcrypto
pytgcalls==1.2.5
yt-dlp
youtube-search-python
